+++ commands/moderation/esay.js
@@ -11,6 +11,11 @@ module.exports = {
     accessableby: 'Staff',
   },
   run: async (bot, message, args, color) => {
+    if (!message.member.guild.me.hasPermission('EMBED_LINKS')) {
+      message.channel.send('I need the permission `Embed Links` for this command!');
+      return;
+    }
+
     if (
       !message.member.hasPermission('MANAGE_GUILD') && message.author.id !== ownerID) {
       message.channel.send(`${language.esay.noPermission}`);
@@ -31,7 +36,9 @@ module.exports = {
       .setColor(color)
       .setDescription(`${sayMessage}`);
 
-    message.delete();
+    if (message.member.guild.me.hasPermission('MANAGE_MESSAGES')) {
+      message.delete();
+    }
 
     message.channel.send(esayEmbed);
   },
