+++ commands/moderation/kick.js
@@ -10,7 +10,9 @@ module.exports = {
     accessableby: 'Staff',
   },
   run: async (bot, message) => {
-    message.delete();
+    if (message.member.guild.me.hasPermission('MANAGE_MESSAGES')) {
+      message.delete();
+    }
 
     // no perms check
