+++ commands/ticket/forceclose.js
@@ -12,7 +12,14 @@ module.exports = {
     accessableby: 'Staff',
   },
   run: async (bot, message, args, color) => {
-    await message.delete();
+    if (!message.member.guild.me.hasPermission('EMBED_LINKS')) {
+      message.channel.send('I need the permission `Embed Links` for this command!');
+      return;
+    }
+
+    if (message.member.guild.me.hasPermission('MANAGE_MESSAGES')) {
+      await message.delete();
+    }
 
     const suppRole = db
       .prepare(
